backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
psycopg2-binary
redis
requests
python-dotenv
aiofiles
opencv-python-headless
numpy
Pillow
